I think ternary add is working.
MAPFUN and CONTAINSP are going to wind up with several different
methods.

For starters, ADD already handles both kinds of sequences.  But, for
KEY-EQUAL to work, the key that we build incrementally as we move
through the tst has to match, or there's no hope of a match.

In turn, MAPFUN ought to also take a sequence type specifier, so it
knows what to pass to its function.  We'll probably wind up with
MAPFUN/LIST and MAPFUN/VECTOR, I imagine.  Or maybe another gf that
specializes on a key type.

Also, we'll want a smarter CONTAINSP.  We know what the next character
in the sequence should be, there's no need to recurse down subtrees
that don't match.  Probably, we can't make CONTAINSP a TCO function,
but that's okay.