Implmented DEFUN-TST macro.
The idea is that a function can be defined via DEFUN-TST to ensure
that :TEST< :TEST= and :IGNORE-CASE keywords are always accepted,
and that TEST< TEST= variables are bound to the proper functions
in the body.

This probably will turn into something that defines standalone
functions, like MAKE, and methods, like ADD.  We want to include
support for the TEST< and TEST= stored in the TST root node, but
allow function invocations to override the defaults set up when
MAKE happened.