Just a last checkin before the do-over.
Lots of stuff appears in this commit.  It's not all related, hell it
might not all even be used one day, but I don't want to lose it right
now.  New structure to the files, packages, and systems on the way,
along with the test framework that works this time.