Changed size/speed, installed better algorithm for splay
Changed one of the *max-vector-sizes* so that :speed 0
is more useful.

Replaced the "count up" algorithm with something a
little more accurate.  We try the square root, cube root,
quad root, and so forth, until we find a depth that
works with the vector limit given by :speed.  I *think*
I like this approach better.