Partially done
Converting over to generating a lambda form when the sparse
vecetor is created, to give the compiler a better shot at
optimizing our spv access.

It works!

I'm checking in now, but the file can't compile because I was
using it like a scratchpad.  Now it's time to clean it up and
add some optional optimizations.