symbols in the wrong package
Worked around for now, but a cleaner solution would get the symbols
referenced in the MAKE-DRILL chain to resolve in the local package.
Instead, I've anchored them all in the package where the forms are
read, which is good enough for now.  But it makes them an eyesore
to read.